debug: introduce debugTelemetry
